requirements-optional.txt
pvporcupine>=3.0,<4.0
webrtcvad>=2.0,<3.0
sounddevice>=0.5,<1.0
numpy>=2.0,<3.0
openai-whisper>=20250625
TTS>=0.22,<1.0
opencv-python>=4.10,<5.0
pytesseract>=0.3,<1.0
Pillow>=11.0,<12.0
PyQt6>=6.8,<7.0
